Intermediate commit: Extensive refactoring and bug fixes
This commit includes a series of fixes across multiple packages to address issues arising from the refactoring of SimulationParameters and other accumulated problems.

Key changes:
- Updated SimParams access in network, neurochemical, pulse, space, and related test files.
- Corrected type mismatches, method calls, and argument lists.
- Resolved undefined constant errors (e.g., PointDimension).
- Fixed syntax errors in cmd package files.
- Cleaned up unused imports and variables in several packages.

Further testing and bug fixing are still required, particularly for test logic in space/spatial_grid_test.go and full validation of cmd package config loading.

+++ pulse/strategy.go
@@ -72,7 +72,7 @@ func (dpp *DefaultPulsePropagator) Propagate(p *Pulse, simParams *config.Simulat
 	if p == nil {
 		return false // Cannot propagate nil pulse
 	}
-	p.CurrentDistance += simParams.General.PulsePropagationSpeed // Assumes SimParams is grouped
+	p.CurrentDistance += float64(simParams.General.PulsePropagationSpeed) // Assumes SimParams is grouped
 	return p.CurrentDistance < p.MaxTravelRadius
 }
 
@@ -91,7 +91,7 @@ func (dpzp *DefaultPulseEffectZoneProvider) GetEffectShell(p *Pulse, simParams *
 		return 0, 0 // Or some other appropriate default for nil pulse
 	}
 	shellEndRadius = p.CurrentDistance
-	shellStartRadius = p.CurrentDistance - simParams.General.PulsePropagationSpeed // Assumes SimParams is grouped
+	shellStartRadius = p.CurrentDistance - float64(simParams.General.PulsePropagationSpeed) // Assumes SimParams is grouped
 	if shellStartRadius < 0 {
 		shellStartRadius = 0
 	}
